converted a lot of ints to longs

+++ src/gwt/org/bbop/apollo/gwt/client/ExonDetailPanel.java
@@ -42,8 +42,8 @@ public class ExonDetailPanel extends Composite {
     interface ExonDetailPanelUiBinder extends UiBinder<Widget, ExonDetailPanel> {
     }
 
-    int inputFmin, inputFmax;
-    int fivePrimeValue, threePrimeValue;
+    Long inputFmin, inputFmax;
+    Long fivePrimeValue, threePrimeValue;
     private AnnotationInfo internalAnnotationInfo;
     private AnnotationInfo annotationInfoWithTopLevelFeature;
     private static ExonDetailPanelUiBinder ourUiBinder = GWT.create(ExonDetailPanelUiBinder.class);
@@ -125,7 +125,7 @@ public class ExonDetailPanel extends Composite {
 
         startColumn = new Column<AnnotationInfo, Number>(new NumberCell()) {
             @Override
-            public Integer getValue(AnnotationInfo annotationInfo) {
+            public Long getValue(AnnotationInfo annotationInfo) {
                 return getDisplayMin(annotationInfo.getMin());
             }
         };
@@ -133,7 +133,7 @@ public class ExonDetailPanel extends Composite {
 
         stopColumn = new Column<AnnotationInfo, Number>(new NumberCell()) {
             @Override
-            public Integer getValue(AnnotationInfo annotationInfo) {
+            public Long getValue(AnnotationInfo annotationInfo) {
                 return annotationInfo.getMax();
             }
         };
@@ -141,7 +141,7 @@ public class ExonDetailPanel extends Composite {
 
         lengthColumn = new Column<AnnotationInfo, Number>(new NumberCell()) {
             @Override
-            public Integer getValue(AnnotationInfo annotationInfo) {
+            public Long getValue(AnnotationInfo annotationInfo) {
                 return annotationInfo.getLength();
             }
         };
@@ -165,21 +165,21 @@ public class ExonDetailPanel extends Composite {
         sortHandler.setComparator(startColumn, new Comparator<AnnotationInfo>() {
             @Override
             public int compare(AnnotationInfo o1, AnnotationInfo o2) {
-                return o1.getMin() - o2.getMin();
+                return o1.getMin().intValue() - o2.getMin().intValue();
             }
         });
 
         sortHandler.setComparator(stopColumn, new Comparator<AnnotationInfo>() {
             @Override
             public int compare(AnnotationInfo o1, AnnotationInfo o2) {
-                return o1.getMax() - o2.getMax();
+                return o1.getMax().intValue() - o2.getMax().intValue();
             }
         });
 
         sortHandler.setComparator(lengthColumn, new Comparator<AnnotationInfo>() {
             @Override
             public int compare(AnnotationInfo o1, AnnotationInfo o2) {
-                return o1.getLength() - o2.getLength();
+                return o1.getLength().intValue() - o2.getLength().intValue();
             }
         });
     }
@@ -398,7 +398,7 @@ public class ExonDetailPanel extends Composite {
         }
     }
     
-    private int getDisplayMin(int min) {
+    private long getDisplayMin(long min) {
         // increases the fmin by 1 for display since coordinates are handled as zero-based on server-side
         return min + 1;
     }
@@ -406,8 +406,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("decreaseFivePrime")
     public void decreaseFivePrimePosition(ClickEvent e) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText()) - 1; // intended action
-            threePrimeValue = Integer.parseInt(threePrimeField.getText());
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText()) - 1; // intended action
+            threePrimeValue = Long.parseLong(threePrimeField.getText());
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -421,8 +421,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("increaseFivePrime")
     public void increaseFivePrimePosition(ClickEvent e) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText()) + 1; // intended action
-            threePrimeValue = Integer.parseInt(threePrimeField.getText());
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText()) + 1; // intended action
+            threePrimeValue = Long.parseLong(threePrimeField.getText());
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -437,8 +437,8 @@ public class ExonDetailPanel extends Composite {
     public void fivePrimeTextEntry(KeyDownEvent k) {
         if (k.getNativeKeyCode() == KeyCodes.KEY_ENTER) {
             try {
-                fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-                threePrimeValue = Integer.parseInt(threePrimeField.getText());
+                fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+                threePrimeValue = Long.parseLong(threePrimeField.getText());
             } catch (Exception error) {
                 coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
                 return;
@@ -455,8 +455,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("fivePrimeField")
     public void fivePrimeTextFocus(BlurEvent b) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-            threePrimeValue = Integer.parseInt(threePrimeField.getText());
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+            threePrimeValue = Long.parseLong(threePrimeField.getText());
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -473,8 +473,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("decreaseThreePrime")
     public void decreaseThreePrimePosition(ClickEvent e) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-            threePrimeValue = Integer.parseInt(threePrimeField.getText()) - 1; // intended action
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+            threePrimeValue = Long.parseLong(threePrimeField.getText()) - 1; // intended action
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -488,8 +488,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("increaseThreePrime")
     public void increaseThreePrime(ClickEvent e) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-            threePrimeValue = Integer.parseInt(threePrimeField.getText()) + 1; // intended action
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+            threePrimeValue = Long.parseLong(threePrimeField.getText()) + 1; // intended action
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -504,8 +504,8 @@ public class ExonDetailPanel extends Composite {
     public void threePrimeTextEntry(KeyDownEvent k) {
         if (k.getNativeKeyCode() == KeyCodes.KEY_ENTER) {
             try {
-                fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-                threePrimeValue = Integer.parseInt(threePrimeField.getText());
+                fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+                threePrimeValue = Long.parseLong(threePrimeField.getText());
             } catch (Exception error) {
                 coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
                 return;
@@ -523,8 +523,8 @@ public class ExonDetailPanel extends Composite {
     @UiHandler("threePrimeField")
     public void threePrimeTextFocus(BlurEvent b) {
         try {
-            fivePrimeValue = Integer.parseInt(fivePrimeField.getText());
-            threePrimeValue = Integer.parseInt(threePrimeField.getText());
+            fivePrimeValue = Long.parseLong(fivePrimeField.getText());
+            threePrimeValue = Long.parseLong(threePrimeField.getText());
         } catch (Exception error) {
             coordinatesToPrime(this.internalAnnotationInfo.getMin(), this.internalAnnotationInfo.getMax());
             return;
@@ -548,16 +548,16 @@ public class ExonDetailPanel extends Composite {
         return true;
     }
     
-    public void triggerUpdate(int fivePrimeValue, int threePrimeValue) {
+    public void triggerUpdate(long fivePrimeValue, long threePrimeValue) {
         final AnnotationInfo originalInfo = this.internalAnnotationInfo;
-        fivePrimeField.setText(Integer.toString(fivePrimeValue));
+        fivePrimeField.setText(Long.toString(fivePrimeValue));
         this.internalAnnotationInfo.setMin(this.inputFmin);
-        threePrimeField.setText(Integer.toString(threePrimeValue));
+        threePrimeField.setText(Long.toString(threePrimeValue));
         this.internalAnnotationInfo.setMax(this.inputFmax);
         updateFeatureLocation(originalInfo);
     }
 
-    public void primeToCoordinates(int fivePrimeFieldValue, int threePrimeFieldValue) {
+    public void primeToCoordinates(long fivePrimeFieldValue, long threePrimeFieldValue) {
         if (this.internalAnnotationInfo.getStrand() == 1) {
             this.inputFmin = fivePrimeFieldValue - 1;
             this.inputFmax = threePrimeFieldValue;
@@ -568,14 +568,14 @@ public class ExonDetailPanel extends Composite {
         }
     }
     
-    public void coordinatesToPrime(int fmin, int fmax) {
+    public void coordinatesToPrime(long fmin, long fmax) {
         if (this.internalAnnotationInfo.getStrand() == 1) {
-            this.fivePrimeField.setText(Integer.toString(fmin + 1));
-            this.threePrimeField.setText(Integer.toString(fmax));
+            this.fivePrimeField.setText(Long.toString(fmin + 1));
+            this.threePrimeField.setText(Long.toString(fmax));
         }
         else {
-            this.fivePrimeField.setText(Integer.toString(fmax));
-            this.threePrimeField.setText(Integer.toString(fmin + 1));
+            this.fivePrimeField.setText(Long.toString(fmax));
+            this.threePrimeField.setText(Long.toString(fmin + 1));
         }
     }
