+++ src/gwt/org/bbop/apollo/gwt/client/rest/SequenceRestService.java
@@ -58,11 +58,11 @@ public class SequenceRestService {
         RestService.sendRequest(requestCallback, "IOService/write", "data=" + jsonObject.toString());
     }
 
-    public static void setCurrentSequenceAndLocation(RequestCallback requestCallback, String sequenceNameString, Integer start, Integer end) {
+    public static void setCurrentSequenceAndLocation(RequestCallback requestCallback, String sequenceNameString, Long start, Long end) {
         setCurrentSequenceAndLocation(requestCallback,sequenceNameString,start,end,false) ;
     }
 
-    public static void setCurrentSequenceAndLocation(RequestCallback requestCallback, String sequenceNameString, Integer start, Integer end,boolean suppressOutput) {
+    public static void setCurrentSequenceAndLocation(RequestCallback requestCallback, String sequenceNameString, Long start, Long end,boolean suppressOutput) {
         String url = "sequence/setCurrentSequenceLocation/?name=" + sequenceNameString + "&start=" + start + "&end=" + end;
         if(suppressOutput){
             url += "&suppressOutput=true";
